Use inline block format also for strings of form "foo\n"
The generated inline strings for snapshots with a single trailing
newline currently look something like this:

```
    insta::assert_snapshot!(stdout, @"foo
");
```

The unindented line there is distracting. I think it looks neater with
the block-style strings, like this:

```
    insta::assert_snapshot!(stdout, @###"
    foo
    "###);
```

I suspect the reason it isn't currently generated like that is just an
oversight caused by the fact that Rust's `str::lines()` returns a
single line for that kind of strings. This patch fixes it by looking
for `\n` characters instead.