Cargo.toml: add crates to workspace
This allows IDE integration to work when the root of the repository is
opened. Prior to this change, a developer would have to open cargo-insta
to get code completion for that crate.

This also causes clippy to run on all crates, so this commit fixes a few
lints.

The integration-tests crate's test is changed to a proper test function
so that it can rely on always being run from the root of the
integration-tests crate. Care is taken to remove the tests directory
after running to avoid pollution in which `cargo test` attempts to run
those tests.