Add the file reference on a YAML error message
I had a process get killed half-way through writing a pending snapshot and it was difficult to debug what was going on, especially since in PRQL we gitignore pending snapshots.

Not sure this is the best way — very open to feedback; in retrospect it seems somewhat verbose to add a simple reference in an error message.