[draft] Upgrade `structopt` to `clap`
Still need to finish this off, one test fails.

`structopt` was build on clap 2, and then its features were integrated into clap 3. So it's not actually a heavier dependency. May have to check on MSRV issues.

Once we merge this, we only have one version of `syn`, which reduces our dependency size significantly.