Don't enforce locking in Cargo.toml
Prepping for release -- these are locked in `Cargo.lock`. Locking them in `Cargo.toml`s could cause a conflict if used as a dependency in combination with other libraries which require a more recent version. This is more of a risk for `insta` rather than `cargo-insta`.

(I'm actually looking forward to the MSRV resolver!)