Unify handling of file & inline snapshots (compat) (#528)
This is a version of #468 that doesn't break compatibility — the change
in snapshot (removing the leading `---` in inline snapshots) is
optional, and snapshots still pass if they have it. Otherwise it has the
same benefits of #468

(There is a tiny corner case that if a snapshot is intended to test
whether there's a leading `---`, we could get a false match, but this is
a very small possibility — much much smaller than removing the line
endings we've done)

As part of that, this adds `as_str_exact` & `matches_fully` methods to
`SnapshotContents`, which allows us to unify our normalization of
snapshots contents, and allow detecting snapshots that pass but don't
completely match; for example if we want to restore checking for line
endings.