draft: Prevent lots of empty inline snapshots with `--force-update-snapshots` without `--accept`
I realize what you mean in the prior comment. Currently running with `--accept` is kinda painful. I'm not sure how common that is, but would be nice to fix.

Had a go for 30 mins, need to run, but pushing an initial effort and a pointer to the thing that needs fixing