Use `INSTA_UPDATE=force` for forcing snapshot updates (#482)
In an effort to simplify the configs, this merges the `INSTA_UPDATE` and
`INSTA_FORCE_UPDATE` configs. Conceptually, `INSTA_FORCE_UPDATE`
overwrites `INSTA_UPDATE`; they naturally fit into the same config
setting.

I realized after starting this that we want to be careful about
supporting new & old versions of `cargo-insta`. So this takes a
conservative approach, ~only changing `insta` at first, but with the
future updates to `cargo-insta` commented in the code. I realize that
adds a bit of complication; though on balance I think simplifying the
configs would be helpful and this makes a step towards that.~ Adjusted
to use the underlying version of `insta`; I think a good approach!

~It stacks on https://github.com/mitsuhiko/insta/pull/479, which should
merge first.~ Merged

~I'd be open to writing some tests for this if that'd be helpful.~
Written