docs: improved the linking inside and outside of the crate (#618)
While browsing the docs on docs.rs, I noticed that there are a few items
which are not linked.
This PR tries to improve this situation.

Drawing attention to one thing which might be controversial:
In the current codebase, the usage of `[´foo´](Self::foo)` and
`[´Self::foo´]` (and related links) are a bit inconsistent. I have
changed all to the `[´Self::foo´]` format. Please let me know if you
prefer the first and I can change them to that format.

I have verified this change via adding the following (somewhat
agressive) docs-related lints and running `RUSTDOCFLAGS="--cfg=docsrs"
cargo +nightly doc --no-deps --all-features` and `cargo clippy
--all-features --all`.

```rust
#![warn(clippy::doc_markdown)]
#![warn(rustdoc::all)]
```

---------

Co-authored-by: Maximilian Roos <5635139+max-sixty@users.noreply.github.com>