Canonicalize snapshot paths before reducing
Fixes https://github.com/mitsuhiko/insta/issues/625

As ever, the reason is a bit different from the one I described there; it's actually that we're reducing the paths by whether one prefixes the other. But in this case the path is only a prefix before it's canonicalized!

On the upside, this case will now work without `--workspace-root`.