Binary snapshots (in-memory) (#610)
The biggest difference to the previous attempt in #489 is that the
snapshot contents are now kept in-memory instead of being written into a
temporary file directly.

For `SnapshotContents` I decided to make it an enum with two variants
like in the previous attempt. I first attempted to change it to three
variants so it would also replace the `SnapshotKind` enum but I realized
that a lot of logic is still shared between inline and file text
snapshots. That approach would definitely also work though and it's
perhaps also a question of how much the logic between the two is
expected to diverge in the future.

In `runtime::assert_snapshot` I replaced the combination of the `refval`
and `new_snapshot_value` with an enum. In that case it is a three
variant enum and some `From` impls for two value tuples for the macros.
Maybe it would be better to make it a two variant enum there too for
consistency and keep the `ReferenceValue` enum for the text case. That
might be a bit weird because both binary and non-inline text snapshot
have a name and that would then be in two different enums.

I should probably also add some tests in `cargo-insta`. I'll do that as
soon as possible, but I need to figure out how the testing there works.

---------

Co-authored-by: Maximilian Roos <5635139+max-sixty@users.noreply.github.com>
Co-authored-by: Maximilian Roos <m@maxroos.com>
Co-authored-by: Frank Elsinga <frank@elsinga.de>