Remove some methods from `SnapshotContents`
Since the binary snapshots have merged, there are some methods which return `Some` if the enum is a certain variant. We can just use pattern matching for that.

(Note that the code is a bit complicated and has lots of repetition of 'kind'-like fields; but wanted to merge the binary snapshots feature before cleaning up the internals