Standardize names for new/old/reference
As discussed in #613.

After doing this, there isn't as much disagreement as I had remembered -- I think we reduced the use of `ReferenceValue` recently as a side-effect of other work. So it's mostly changing names from `old` & `new` to `existing` & `generated`/`proposed`.

Opening the PR but will leave for a few days to consider whether it's worthwhile