[ISSUE-718] Allow specifying cargo-nextest bin (#721)
Should fix #718 

In some environments, `cargo-nextest` is not installed where `cargo`
expects it. (e.g. in nix environment).

This MR aims to  be able to explicitly specify `cargo-nextest` bin.


First MR in this repository, I might not be fully aware of the good
practices, happy to take reviews on this :)

---------

Co-authored-by: Louis Fruleux <louis.fruleux@helsing.ai>
Co-authored-by: Maximilian Roos <5635139+max-sixty@users.noreply.github.com>