Fix test when not in `insta` path (#744)
The existing test expects to be in a path that ends with `insta`. In a
git worktree that's often not the case.

I think this change may make the test a bit impotent, but it does fix
the issue, and I don't think the test is _that_ important. (other ideas
welcome as ever)